refactor(sensors): address PR feedback on Phenomenon model
- Move phenomenon endpoints out of ReferenceDataEndpoints into their own
  PhenomenonEndpoints file; register via SensorsApiExtensions.
- Switch IPhenomenonRepository.GetAllAsync / GetByCapabilityAsync to
  IAsyncEnumerable to match the streaming pattern used elsewhere in the
  sensors data-access layer (ISensorRepository, IReadingRepository,
  ISensorHealthRepository). Singular getters keep their Task<T?> shape.
- Drop the explicit AsNoTracking call in ParameterResolver — the
  SensorsDbContext is already configured with QueryTrackingBehavior.NoTracking
  as the default in SensorsDatabaseExtensions.
- Remove unused single-call ParameterResolver.ResolveAsync; only the batch
  LoadLookupAsync path is consumed today (by UsgsIngestionWorker), and the
  in-memory ParameterLookup.Resolve already covers the single-reading case
  whenever a future Push API path needs it.
